requirements.txt
streamlit==1.26.0
pymongo[srv]
python-dotenv==1.0.0